==> 01_data_collection_colorectal.ipynb <==
import pandas as pd

# merge patient record and their sample id
patient = pd.read_csv('coadread_tcga/data_clinical_patient.txt', delimiter='\t', skiprows=4)
patient.dropna(axis=0, subset=['OS_MONTHS'], inplace=True)
sample = pd.read_csv('coadread_tcga/data_clinical_sample.txt', delimiter='\t', skiprows=4)
clinical_merged = pd.merge(sample, patient, on='PATIENT_ID')
clinical_merged

# process mrna data
mrna_df = pd.read_csv('coadread_tcga/data_linear_cna.txt', delimiter='\t')
mrna = mrna_df.dropna()
mrna = mrna.transpose()
mrna.drop('Entrez_Gene_Id', axis=0, inplace=True)
mrna.reset_index(inplace=True)
mrna.columns = mrna.iloc[0]
mrna = mrna[1:].reset_index(drop=True)
mrna.rename(columns={'Hugo_Symbol': 'SAMPLE_ID'}, inplace=True)

mrna

# merge clinical data and mrna data, drop na values
dataset = pd.merge(mrna, clinical_merged[['PATIENT_ID', 'SAMPLE_ID', 'OS_STATUS', 'OS_MONTHS', 'DFS_STATUS', 'DFS_MONTHS']], on='SAMPLE_ID')
dataset.replace('[Not Available]', pd.NA, inplace=True)
dataset.dropna(subset=['OS_STATUS'], inplace=True)
dataset['OS_STATUS'] = dataset['OS_STATUS'].map({'0:LIVING': 0, '1:DECEASED': 1})
dataset.to_csv('dataset.csv', index=False)
